pyproject.toml
[tool.poetry]
name = "airport-api"
version = "1.0.0"
description = "Airport API service"
authors = ["Vladyslav Timofieiev <vladislav.tmf@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-filter = "^24.3"
django-redis = "^5.4.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^11.0.0"
redis = "^5.2.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"
black = "^24.10.0"
django-jazzmin = "^3.0.1"
drf-spectacular = "^0.28.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
